fix/booking

+++ database/migrations/2024_10_10_125856_create_theater_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('theater', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('id_place_theater'); 
+            $table->integer('id_place_theater'); 
             $table->string('name', 255); 
             $table->string('descript', 255)->nullable();
             $table->string('address', 255);
